+++ genomics_data_index/pipelines/PipelineExecutor.py
@@ -1,11 +1,14 @@
 import abc
+import gzip
 import io
 import logging
+import random
 import sys
 from pathlib import Path
-from typing import List, Union, Tuple, Optional
+from typing import List, Union, Tuple, Optional, Set
 
 import pandas as pd
+from Bio import SeqIO
 from pathvalidate import sanitize_filename
 
 from genomics_data_index.pipelines.ExecutorResults import ExecutorResults
@@ -189,6 +192,73 @@ class PipelineExecutor(abc.ABC):
             else:
                 input_sample_files[col] = input_sample_files[col].apply(lambda x: str(x) if not pd.isna(x) else pd.NA)
         input_sample_files.to_csv(output_file, sep='\t', index=False)
+
+    def select_random_samples(self, input_sequence_files: List[Path],
+                              number_samples: float, random_seed: int = None) -> Set[str]:
+        """
+        Selects a random set of samples from the passed sequence files (assumes one sample is one sequence record).
+        :param input_sequence_files: The list of sequence files to select samples from.
+        :param number_samples: If >= 1, then represents the number of samples to select (e.g., 5 for 5 samples).
+                               If < 1, then represents the proportion of samples in all files to select
+                               (e.g., 0.5 means 50% of samples across all files).
+        :param random_seed: The random seed. If None uses the default Python random seed
+                            (system time or some other method, see Python documentation on random.seed()).
+        :return: A randomly selected set of sample names from the passed sequence files (no check is performed
+                 for duplicate sequences across files).
+        """
+        if number_samples < 0:
+            raise Exception(f'number_samples={number_samples} must be non-negative')
+
+        samples = []
+        for input_sequence_file in input_sequence_files:
+            sequence_file = SequenceFile(input_sequence_file)
+            for record in sequence_file.records():
+                samples.append(record.id)
+
+        if number_samples < 1:
+            number_samples = round(number_samples * len(samples))
+        else:
+            number_samples = round(number_samples)
+
+        random.seed(random_seed)
+        return set(random.sample(samples, k=number_samples))
+
+    def split_input_sequence_files(self, input_sequence_files: List[Path], output_dir: Path,
+                                   samples: Set[str] = None) -> pd.DataFrame:
+        if samples is not None and isinstance(samples, list):
+            samples = set(samples)
+
+        sample_data = []
+        print_on = 200
+        for input_sequence_file in input_sequence_files:
+            count = 0
+            sequence_file = SequenceFile(input_sequence_file)
+
+            if sequence_file.is_fasta():
+                extension = '.fasta.gz'
+                filetype = 'fasta'
+            elif sequence_file.is_genbank():
+                extension = '.gbk.gz'
+                filetype = 'genbank'
+            else:
+                raise Exception(f'Unknown filetype for {input_sequence_file}, must be either fasta or genbank')
+
+            for record in sequence_file.records():
+                if count % print_on == 0:
+                    logger.debug(f'Processed {count} sequences from {input_sequence_file}')
+
+                sample_name = record.id
+
+                if samples is None or sample_name in samples:
+                    sample_filename = sanitize_filename(sample_name, replacement_text='_')
+                    sample_path = output_dir / (sample_filename + extension)
+                    with gzip.open(sample_path, "wt") as oh:
+                        SeqIO.write(record, oh, filetype)
+
+                    sample_data.append([sample_name, sample_path, pd.NA, pd.NA])
+
+                count += 1
+        return pd.DataFrame(sample_data, columns=self.INPUT_SAMPLE_FILE_COLUMNS)
 
     def read_input_sample_files(self, input_file: Path) -> pd.DataFrame:
         df = pd.read_csv(input_file, sep='\t')